requirements/base.txt
# Core Django
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
python-decouple>=3.8

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
django-filter>=23.5

# Django Channels for WebSockets
channels>=4.0.0
channels-redis>=4.2.0
daphne>=4.1.0

# Social Authentication
django-allauth>=0.57.0
django-allauth[socialaccount]>=0.57.0

# 2FA Authentication
django-two-factor-auth>=1.15.2

# File Storage
django-storages>=1.14.2
boto3>=1.34.0

# Image Processing
Pillow>=10.2.0

# PDF Generation
reportlab>=4.0.9

# Payments
stripe>=8.0.0

# Email
django-anymail>=10.2

# Utilities
python-slugify>=8.0.1
pytz>=2024.1

# Security
django-ratelimit>=4.1.0
